+++ odoo/songs/install/accounting.py
@@ -207,10 +207,30 @@ def add_xmlid_tax(ctx):
                         company_xmlid_name,
                         tax_name,
                     ),
-                'module': 'scenario',
+                'module': '__setup__',
                 'model': 'account.tax',
                 'res_id': tax.id
             })
+
+
+@anthem.log
+def add_xmlid_fiscal_position(ctx):
+    fiscal_positions = ctx.env['account.fiscal.position'].search([])
+    for pos in fiscal_positions:
+        if 'Extra' in pos.name:
+            code = 'extra'
+        elif 'Intra' in pos.name:
+            code = 'intra'
+        elif 'National' in pos.name:
+            code = 'nat'
+        else:
+            code = 'cocontractor'
+
+        add_xmlid(
+            ctx, pos,
+            '__setup__.fiscal_position_' + code,
+            noupdate=True
+            )
 
 
 @anthem.log
@@ -264,6 +284,7 @@ def main(ctx):
     create_financial_journals(ctx)
     add_xmlid_account(ctx)
     add_xmlid_tax(ctx)
+    add_xmlid_fiscal_position(ctx)
     adapt_chart_of_account(ctx)
     settings(ctx)
     default_values(ctx)