+++ odoo/songs/install/data_full.py
@@ -67,21 +67,29 @@ def import_pricelist_items(ctx):
 
 
 @anthem.log
-def import_locators(ctx):
-    """ Importing locators from csv"""
-    content = resource_stream(req, 'data/demo/locators_subset.csv')
-    load_csv_stream(ctx, 'stock.location', content, delimiter=',')
-    content = resource_stream(req, 'data/demo/locators_reserve.csv')
-    load_csv_stream(ctx, 'stock.location', content, delimiter=',')
-    content = resource_stream(req, 'data/demo/locators_parking.csv')
-    load_csv_stream(ctx, 'stock.location', content, delimiter=',')
+def import_locations(ctx):
+    """ Importing locations from csv"""
 
+    load_ctx = ctx.env.context.copy()
+    load_ctx.update({'defer_parent_store_computation': True})
+    Location = ctx.env['stock.location'].with_context(load_ctx)
+    with ctx.log(u"Importing warehouse locations"):
+        content = resource_stream(req, 'data/install/location.csv')
+        load_csv_stream(ctx, Location, content, delimiter=',',
+                        context=load_ctx)
+    with ctx.log(u"Importing reserve locations"):
+        content = resource_stream(req, 'data/demo/locators_reserve.csv')
+        load_csv_stream(ctx, 'stock.location', content, delimiter=',')
+    with ctx.log(u"Importing parking locations"):
+        content = resource_stream(req, 'data/demo/locators_parking.csv')
+        load_csv_stream(ctx, 'stock.location', content, delimiter=',')
+    with ctx.log(u"Importing output locations"):
+        content = resource_stream(req, 'data/install/chariots.csv')
+        load_csv_stream(ctx, Location, content, delimiter=',',
+                        context=load_ctx)
 
-@anthem.log
-def import_output_locations(ctx):
-    """ Importing output locations from csv"""
-    content = resource_stream(req, 'data/demo/chariots.csv')
-    load_csv_stream(ctx, 'stock.location', content, delimiter=',')
+    with ctx.log(u"Compute parent_left, parent_right"):
+        ctx.env['stock.location']._parent_store_compute()
 
 
 @anthem.log
@@ -101,6 +109,5 @@ def main(ctx):
     other files will be import by importer.sh)
     """
     # Putting some demo data in full mode because we don't have yet real data
-    import_locators(ctx)
-    import_output_locations(ctx)
+    import_locations(ctx)
     import_delivery_round_config(ctx)